refactor: remove shared barrel imports
Replace internal shared/index.js imports with imports from the defining shared modules. Add a Biome restricted-import rule so the shared barrel can remain available without becoming an internal dependency shortcut.
