+++ biome.json
@@ -27,6 +27,10 @@
               {
                 "group": ["**/services/index.js"],
                 "message": "Import from the defining service module instead of the services barrel."
+              },
+              {
+                "group": ["**/shared/index.js"],
+                "message": "Import from the defining shared module instead of the shared barrel."
               }
             ]
           }
