+++ src/commands/code/files.ts
@@ -7,12 +7,6 @@ import {
 } from "../../shared/code-navigation-defaults.js";
 import { shouldUseColors } from "../../shared/colors.js";
 import {
-  InvalidPackageSpecError,
-  requireAuth,
-  SPINNER_MESSAGES,
-  startSpinner,
-} from "../../shared/index.js";
-import {
   buildListFilesParams,
   type ListFilesRequestBuildResult,
   type ListFilesRequestInput,
@@ -21,11 +15,15 @@ import {
   buildListFilesSuccessPayload,
   formatListFilesTerminal,
 } from "../../shared/list-files-response.js";
+import { InvalidPackageSpecError } from "../../shared/package-spec.js";
 import {
   PKGSEER_REGISTRY_ARGS,
   PKGSEER_REGISTRY_LIST,
   toPkgseerRegistryLowercase,
 } from "../../shared/pkgseer-registry.js";
+import { requireAuth } from "../../shared/require-auth.js";
+import { startSpinner } from "../../shared/spinner.js";
+import { SPINNER_MESSAGES } from "../../shared/spinner-messages.js";
 import {
   formatIndexingError,
   handleCodeNavCommandError,
