+++ src/commands/code/grep.ts
@@ -7,18 +7,20 @@ import {
 import { shouldUseColors } from "../../shared/colors.js";
 import {
   buildGrepRepoParams,
-  buildGrepRepoSuccessPayload,
-  formatGrepRepoTerminal,
   GREP_REPO_PATTERN_NOTE,
   GREP_REPO_SYMBOL_FIELDS_NOTE,
   type GrepRepoRequestBuildResult,
   type GrepRepoRequestInput,
-  InvalidPackageSpecError,
-  requireAuth,
-  SPINNER_MESSAGES,
-  startSpinner,
-} from "../../shared/index.js";
+} from "../../shared/grep-repo-request.js";
+import {
+  buildGrepRepoSuccessPayload,
+  formatGrepRepoTerminal,
+} from "../../shared/grep-repo-response.js";
+import { InvalidPackageSpecError } from "../../shared/package-spec.js";
 import { toPkgseerRegistryLowercase } from "../../shared/pkgseer-registry.js";
+import { requireAuth } from "../../shared/require-auth.js";
+import { startSpinner } from "../../shared/spinner.js";
+import { SPINNER_MESSAGES } from "../../shared/spinner-messages.js";
 import {
   formatFileErrorWithFilesHint,
   handleCodeNavCommandError,
