+++ src/commands/docs/read.ts
@@ -1,18 +1,18 @@
 import type { Command } from "commander";
 import { createContainer } from "../../container.js";
 import type { PackageIntelligenceService } from "../../services/package-intelligence-service.js";
+import { shouldUseColors } from "../../shared/colors.js";
+import { mapPackageIntelligenceError } from "../../shared/package-intelligence-error-map.js";
+import { InvalidPackageSpecError } from "../../shared/package-spec.js";
+import { parseLinesOption } from "../../shared/parse-lines-option.js";
+import { buildReadPackageDocParams } from "../../shared/read-package-doc-request.js";
 import {
-  buildReadPackageDocParams,
   buildReadPackageDocSuccessPayload,
   formatReadPackageDocTerminal,
-  InvalidPackageSpecError,
-  mapPackageIntelligenceError,
-  parseLinesOption,
-  requireAuth,
-  SPINNER_MESSAGES,
-  shouldUseColors,
-  startSpinner,
-} from "../../shared/index.js";
+} from "../../shared/read-package-doc-response.js";
+import { requireAuth } from "../../shared/require-auth.js";
+import { startSpinner } from "../../shared/spinner.js";
+import { SPINNER_MESSAGES } from "../../shared/spinner-messages.js";
 import {
   buildCliMappedErrorPayload,
   formatMappedErrorForTerminal,
