+++ src/commands/pkg/changelog.ts
@@ -2,22 +2,24 @@ import type { Command } from "commander";
 import { createContainer } from "../../container.js";
 import type { PackageIntelligenceService } from "../../services/package-intelligence-service.js";
 import { shouldUseColors } from "../../shared/colors.js";
-import {
-  InvalidPackageSpecError,
-  type MappedError,
-  mapPackageIntelligenceError,
-  parsePackageSpec,
-  requireAuth,
-} from "../../shared/index.js";
 import { buildPackageChangelogParams } from "../../shared/package-changelog-request.js";
 import {
   buildPackageChangelogSuccessPayload,
   formatPackageChangelogTerminal,
 } from "../../shared/package-changelog-response.js";
 import {
+  type MappedError,
+  mapPackageIntelligenceError,
+} from "../../shared/package-intelligence-error-map.js";
+import {
+  InvalidPackageSpecError,
+  parsePackageSpec,
+} from "../../shared/package-spec.js";
+import {
   PKGSEER_REGISTRY_LIST,
   toPkgseerRegistryLowercase,
 } from "../../shared/pkgseer-registry.js";
+import { requireAuth } from "../../shared/require-auth.js";
 import {
   buildCliMappedErrorPayload,
   formatMappedErrorForTerminal,
