+++ src/commands/pkg/info.ts
@@ -2,18 +2,18 @@ import type { Command } from "commander";
 import { createContainer } from "../../container.js";
 import type { PackageIntelligenceService } from "../../services/package-intelligence-service.js";
 import { shouldUseColors } from "../../shared/colors.js";
+import { mapPackageIntelligenceError } from "../../shared/package-intelligence-error-map.js";
 import {
   InvalidPackageSpecError,
-  mapPackageIntelligenceError,
   parsePackageSpec,
-  requireAuth,
-} from "../../shared/index.js";
+} from "../../shared/package-spec.js";
 import { buildPackageSummaryParams } from "../../shared/package-summary-request.js";
 import {
   buildPackageSummarySuccessPayload,
   formatPackageSummaryTerminal,
 } from "../../shared/package-summary-response.js";
 import { PKGSEER_REGISTRY_LIST } from "../../shared/pkgseer-registry.js";
+import { requireAuth } from "../../shared/require-auth.js";
 import {
   buildCliMappedErrorPayload,
   formatMappedErrorForTerminal,
